requirements.txt
streamlit==1.47.0
faiss-cpu
google-generativeai==0.7.2
sentence-transformers==2.7.0
pandas==2.3.1
numpy==2.3.1
tqdm==4.67.1
requests==2.32.4
beautifulsoup4==4.13.4
